requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bitarray==3.3.1
ccxt==3.0.71
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
ckzg==2.1.1
comm==0.1.3
constantly==15.1.0
cryptocompy==0.1.1.dev1
cryptography==40.0.2
cytoolz==1.0.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hexbytes==1.3.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.5.0
incremental==22.10.0
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
keyring==23.13.1
MarkupSafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
nest-asyncio==1.5.6
okex==0.1.6
packaging==23.1
parsimonious==0.10.0
parso==0.8.3
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.3.0
pycparser==2.21
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.15.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.7
python-okx==0.1.9
pyunormalize==16.0.0
PyYAML==6.0.2
pyzmq==24.0.1
regex==2024.11.6
rel==0.4.8
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
toolz==1.0.0
tornado==6.2
Twisted==22.10.0
txaio==23.1.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspection==0.4.0
typing_extensions==4.13.2
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
web3==7.10.0
websocket-client==1.5.1
websockets==11.0.2
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
